Additional cleaning up
