+++ frontend/src/app/elections/[id]/candidates/page.tsx
@@ -5,11 +5,10 @@ export const metadata: Metadata = {
   title: "ABVS | Election Candidates",
 }
 
-export default function ElectionCandidates() {
+export default function ElectionCandidatesPage() {
   return (
     <div>
       <AddCandidate />
-      <div>List or Table of Already Added Candidates</div>
     </div>
   );
 }
