+++ frontend/src/app/elections/[id]/verify/page.tsx
@@ -5,7 +5,7 @@ export const metadata: Metadata = {
   title: "ABVS | Election Verify",
 };
 
-export default function ElectionVerify() {
+export default function ElectionVerifyPage() {
   return (
     <main className="w-full flex items-center justify-center mb-40">
       <VerifyVoter />
