+++ frontend/src/app/elections/page.tsx
@@ -5,7 +5,7 @@ export const metadata: Metadata = {
   title: "ABVS | Elections",
 }
 
-export default function Elections() {
+export default function ElectionsPage() {
   return (
     <div className="p-4 md:p-6 lg:p-8 mb-32 lg:mb-0">
       <div className="flex flex-col items-center text-center mb-6">
